refactor(projects): extract line-item edit payload logic into a shared module
Pulls the discount-entry seeding, form-defaults, and $/% payload resolution
out of EditLineItemDialog's handleSave/defaultValues into
src/lib/line-item-edit-payload.ts, with no behavior change. Sets up a single
source of truth the upcoming inline-edit cells can reuse instead of
hand-duplicating the discount math (POLICY.md R-3.1).
